refac: move components to new folders to follow atomic design
