+++ src/components/pages/register/page.tsx
@@ -1,4 +1,4 @@
-import RegisterForm from "@/components/auth/RegisterForm";
+import RegisterForm from "@/components/organisms/Registerform/RegisterForm";
 import React from "react";
 
 export default function page() {
